requirements.txt
altgraph==0.17.4
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
coverage==7.10.2
decorator==5.2.1
future==1.0.0
idna==3.10
packaging==25.0
pyswisseph==2.10.3.2
ratelim==0.1.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
urllib3==2.5.0
